// Week10/gureum/mission1/src/types/movie.ts
export type MovieListType = 'popular' | 'upcoming' | 'top_rated';
export type MovieLanguage = 'ko-KR' | 'en-US' | 'ja-JP';

export interface MovieSearchParams {
  query: string;
  includeAdult: boolean;
  language: MovieLanguage;
  page: number;
}

export interface Genre {
  id: number;
  name: string;
}

export interface Movie {
  id: number;
  title: string;
  overview: string;
  poster_path: string;
  release_date: string;
  vote_average: number;
}

export interface MovieListResponse {
  page: number;
  results: Movie[];
  total_pages: number;
  total_results: number;
}

export interface MovieDetail {
  id: number;
  title: string;
  overview: string;
  poster_path: string;
  backdrop_path: string;
  release_date: string;
  vote_average: number;
  vote_count: number;
  runtime: number;
  genres: Genre[];
  budget: number;
  revenue: number;
  tagline: string;
  homepage: string;
  status: string;
  original_language: string;
  original_title: string;
  popularity: number;
}

export interface CastMember {
  id: number;
  name: string;
  character: string;
  profile_path: string | null;
}

export interface CrewMember {
  id: number;
  name: string;
  job: string;
}

export interface Credits {
  cast: CastMember[];
  crew: CrewMember[];
}